Cart Controller and Service modifications
CartController:
- Added @PathVariable and a new GET mapping for adding a product to the cart
- Updated the getAllCarts method to use the CartService to find all carts

+++ src/main/java/com/example/ecomweb/Entity/UserBean.java
@@ -17,7 +17,7 @@ import lombok.*;
 public class UserBean {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
+    private Long id;
     private String name, address;
-    private long phoneNumber;    
+    private Long phoneNumber;    
 }
